Add genpybind-generated wrapping
Depends-On: 11824

+++ pyhwdb/test/pyhwdb_test.py
@@ -3,29 +3,39 @@
 import unittest
 import os
 import pyhwdb
-import pyhalco_hicann_v2 as coord
+
+IS_PYPLUSPLUS = None
+try:
+    import pyhalco_hicann_v2 as coord
+    IS_PYPLUSPLUS = True
+except ImportError:
+    import pyhalco_common as coord
+    IS_PYPLUSPLUS = False
 
 class Test_Pyhwdb(unittest.TestCase):
 
     def __init__(self, *args, **kwargs):
         super(Test_Pyhwdb, self).__init__(*args, **kwargs)
-        self.WAFER_COORD = coord.Wafer(10)
-        self.WAFER_SETUP_TYPE = coord.SetupType.BSSWafer
-        self.WAFER_MACU_IP = coord.IPv4.from_string("192.168.10.120")
-        self.FPGA_COORD = coord.FPGAOnWafer(0)
-        self.FPGA_HIGHSPEED = True
+        if IS_PYPLUSPLUS:
+            self.WAFER_COORD = coord.Wafer(10)
+            self.WAFER_SETUP_TYPE = coord.SetupType.BSSWafer
+            self.WAFER_MACU_IP = coord.IPv4.from_string("192.168.10.120")
+            self.FPGA_COORD = coord.FPGAOnWafer(0)
+            self.FPGA_HIGHSPEED = True
         self.FPGA_IP = coord.IPv4.from_string("192.168.10.1")
-        self.RETICLE_COORD = coord.DNCOnWafer(coord.common.Enum(0))
-        self.RETICLE_TO_BE_POWERED = False
-        self.HICANN_COORD = coord.HICANNOnWafer(coord.common.Enum(144))
-        self.HICANN_VERSION = 4
-        self.HICANN_LABEL = "bla"
-        self.ADC_COORD = "B123456"
-        self.ADC_MODE = "LOAD_CALIBRATION"
-        self.ADC_CHAN = coord.ChannelOnADC(0)
-        self.ADC_TRIGGER = coord.TriggerOnADC(1)
+        if IS_PYPLUSPLUS:
+            self.RETICLE_COORD = coord.DNCOnWafer(coord.common.Enum(0))
+            self.RETICLE_TO_BE_POWERED = False
+            self.HICANN_COORD = coord.HICANNOnWafer(coord.common.Enum(144))
+            self.HICANN_VERSION = 4
+            self.HICANN_LABEL = "bla"
+            self.ADC_COORD = "B123456"
+            self.ADC_MODE = "LOAD_CALIBRATION"
+            self.ADC_CHAN = coord.ChannelOnADC(0)
+            self.ADC_TRIGGER = coord.TriggerOnADC(1)
         self.ADC_IP = coord.IPv4.from_string("192.168.10.120")
-        self.ADC_PORT = coord.TCPPort(10101)
+        if IS_PYPLUSPLUS:
+            self.ADC_PORT = coord.TCPPort(10101)
 
         self.DLS_SETUP_ID = "07_20"
         self.HXCUBE_ID = 9
@@ -51,6 +61,7 @@ class Test_Pyhwdb(unittest.TestCase):
         self.assertTrue(os.path.exists(default_path))
         db.load(default_path)
 
+    @unittest.skipUnless(IS_PYPLUSPLUS, "Only works for wafer currently")
     def test_add_to_empty_db(self):
         mydb = pyhwdb.database()
         fpga_coord = coord.FPGAGlobal(self.FPGA_COORD, self.WAFER_COORD)
@@ -69,17 +80,18 @@ class Test_Pyhwdb(unittest.TestCase):
     def test_database_access(self):
         mydb = pyhwdb.database()
 
-        wafer = pyhwdb.WaferEntry()
-        wafer_coord = self.WAFER_COORD
-        wafer.setup_type = self.WAFER_SETUP_TYPE
-        self.assertFalse(mydb.has_wafer_entry(wafer_coord))
-        mydb.add_wafer_entry(wafer_coord, wafer)
-        self.assertTrue(mydb.has_wafer_entry(wafer_coord))
-        self.assertEqual(mydb.get_wafer_entry(wafer_coord).setup_type, wafer.setup_type)
-        wafer_coords = mydb.get_wafer_coordinates()
-        self.assertEqual(len(wafer_coords), 1)
-        mydb.remove_wafer_entry(wafer_coord)
-        self.assertFalse(mydb.has_wafer_entry(wafer_coord))
+        if IS_PYPLUSPLUS:
+            wafer = pyhwdb.WaferEntry()
+            wafer_coord = self.WAFER_COORD
+            wafer.setup_type = self.WAFER_SETUP_TYPE
+            self.assertFalse(mydb.has_wafer_entry(wafer_coord))
+            mydb.add_wafer_entry(wafer_coord, wafer)
+            self.assertTrue(mydb.has_wafer_entry(wafer_coord))
+            self.assertEqual(mydb.get_wafer_entry(wafer_coord).setup_type, wafer.setup_type)
+            wafer_coords = mydb.get_wafer_coordinates()
+            self.assertEqual(len(wafer_coords), 1)
+            mydb.remove_wafer_entry(wafer_coord)
+            self.assertFalse(mydb.has_wafer_entry(wafer_coord))
 
         dls_entry = pyhwdb.DLSSetupEntry()
         dls_setup_id = self.DLS_SETUP_ID
@@ -104,62 +116,63 @@ class Test_Pyhwdb(unittest.TestCase):
         self.assertTrue(mydb.has_hxcube_entry(hxcube_id))
         mydb.remove_hxcube_entry(hxcube_id)
         self.assertFalse(mydb.has_hxcube_entry(hxcube_id))
-        
-
-        # require wafer entry to write other entry typed into
-        mydb.add_wafer_entry(wafer_coord, wafer)
-
-        fpga = pyhwdb.FPGAEntry()
-        fpga_coord = coord.FPGAGlobal(self.FPGA_COORD, self.WAFER_COORD)
-        fpga.ip = self.FPGA_IP
-        self.assertFalse(mydb.has_fpga_entry(fpga_coord))
-        mydb.add_fpga_entry(fpga_coord, fpga)
-        self.assertTrue(mydb.has_fpga_entry(fpga_coord))
-        self.assertEqual(mydb.get_fpga_entry(fpga_coord).ip, fpga.ip)
-        mydb.remove_fpga_entry(fpga_coord)
-        self.assertFalse(mydb.has_fpga_entry(fpga_coord))
-
-        reticle = pyhwdb.ReticleEntry()
-        reticle_coord = coord.DNCGlobal(self.RETICLE_COORD, self.WAFER_COORD)
-        reticle.to_be_powered = self.RETICLE_TO_BE_POWERED
-        self.assertFalse(mydb.has_reticle_entry(reticle_coord))
-        mydb.add_reticle_entry(reticle_coord, reticle)
-        self.assertTrue(mydb.has_reticle_entry(reticle_coord))
-        self.assertEqual(mydb.get_reticle_entry(reticle_coord).to_be_powered, reticle.to_be_powered)
-        mydb.remove_reticle_entry(reticle_coord)
-        self.assertFalse(mydb.has_reticle_entry(reticle_coord))
 
 
-        # require fpga entry to add hicann entry
-        mydb.add_fpga_entry(fpga_coord, fpga)
+        if IS_PYPLUSPLUS:
+            # require wafer entry to write other entry typed into
+            mydb.add_wafer_entry(wafer_coord, wafer)
 
-        hicann = pyhwdb.HICANNEntry()
-        hicann_coord = coord.HICANNGlobal(self.HICANN_COORD, self.WAFER_COORD)
-        hicann.version = self.HICANN_VERSION
-        self.assertFalse(mydb.has_hicann_entry(hicann_coord))
-        mydb.add_hicann_entry(hicann_coord, hicann)
-        self.assertTrue(mydb.has_hicann_entry(hicann_coord))
-        self.assertEqual(mydb.get_hicann_entry(hicann_coord).version, hicann.version)
-        mydb.remove_hicann_entry(hicann_coord)
-        self.assertFalse(mydb.has_hicann_entry(hicann_coord))
+            fpga = pyhwdb.FPGAEntry()
+            fpga_coord = coord.FPGAGlobal(self.FPGA_COORD, self.WAFER_COORD)
+            fpga.ip = self.FPGA_IP
+            self.assertFalse(mydb.has_fpga_entry(fpga_coord))
+            mydb.add_fpga_entry(fpga_coord, fpga)
+            self.assertTrue(mydb.has_fpga_entry(fpga_coord))
+            self.assertEqual(mydb.get_fpga_entry(fpga_coord).ip, fpga.ip)
+            mydb.remove_fpga_entry(fpga_coord)
+            self.assertFalse(mydb.has_fpga_entry(fpga_coord))
 
-        # test clear()
-        mydb.add_hicann_entry(hicann_coord, hicann)
-        mydb.clear()
-        self.assertFalse(mydb.has_hicann_entry(hicann_coord))
-        self.assertFalse(mydb.has_fpga_entry(fpga_coord))
-        self.assertFalse(mydb.has_wafer_entry(wafer_coord))
+            reticle = pyhwdb.ReticleEntry()
+            reticle_coord = coord.DNCGlobal(self.RETICLE_COORD, self.WAFER_COORD)
+            reticle.to_be_powered = self.RETICLE_TO_BE_POWERED
+            self.assertFalse(mydb.has_reticle_entry(reticle_coord))
+            mydb.add_reticle_entry(reticle_coord, reticle)
+            self.assertTrue(mydb.has_reticle_entry(reticle_coord))
+            self.assertEqual(mydb.get_reticle_entry(reticle_coord).to_be_powered, reticle.to_be_powered)
+            mydb.remove_reticle_entry(reticle_coord)
+            self.assertFalse(mydb.has_reticle_entry(reticle_coord))
 
-        # FIXME: get usable ADC coord, i.e. proper wrapping
-        #adc = pyhwdb.ADCEntry()
-        #adc_coord = coord.ADCGlobal(self.ADC_COORD, self.WAFER_COORD)
-        #adc.version = self.ADC_CAHNNEL
-        #self.assertFalse(mydb.has_adc_entry(adc_coord))
-        #mydb.add_adc_entry(adc_coord, adc)
-        #self.assertTrue(mydb.has_adc_entry(adc_coord))
-        #self.assertEqual(mydb.get_adc_entry(adc_coord).channel, adc.channel)
-        #mydb.remove_adc_entry(adc_coord)
-        #self.assertFalse(mydb.has_adc_entry(adc_coord))
+
+            # require fpga entry to add hicann entry
+            mydb.add_fpga_entry(fpga_coord, fpga)
+
+            hicann = pyhwdb.HICANNEntry()
+            hicann_coord = coord.HICANNGlobal(self.HICANN_COORD, self.WAFER_COORD)
+            hicann.version = self.HICANN_VERSION
+            self.assertFalse(mydb.has_hicann_entry(hicann_coord))
+            mydb.add_hicann_entry(hicann_coord, hicann)
+            self.assertTrue(mydb.has_hicann_entry(hicann_coord))
+            self.assertEqual(mydb.get_hicann_entry(hicann_coord).version, hicann.version)
+            mydb.remove_hicann_entry(hicann_coord)
+            self.assertFalse(mydb.has_hicann_entry(hicann_coord))
+
+            # test clear()
+            mydb.add_hicann_entry(hicann_coord, hicann)
+            mydb.clear()
+            self.assertFalse(mydb.has_hicann_entry(hicann_coord))
+            self.assertFalse(mydb.has_fpga_entry(fpga_coord))
+            self.assertFalse(mydb.has_wafer_entry(wafer_coord))
+
+            # FIXME: get usable ADC coord, i.e. proper wrapping
+            #adc = pyhwdb.ADCEntry()
+            #adc_coord = coord.ADCGlobal(self.ADC_COORD, self.WAFER_COORD)
+            #adc.version = self.ADC_CAHNNEL
+            #self.assertFalse(mydb.has_adc_entry(adc_coord))
+            #mydb.add_adc_entry(adc_coord, adc)
+            #self.assertTrue(mydb.has_adc_entry(adc_coord))
+            #self.assertEqual(mydb.get_adc_entry(adc_coord).channel, adc.channel)
+            #mydb.remove_adc_entry(adc_coord)
+            #self.assertFalse(mydb.has_adc_entry(adc_coord))
 
 
 if __name__ == "__main__":
